+++ src/MyServiceBus/ITransportFactory.cs
@@ -6,6 +6,11 @@ public interface ITransportFactory
 {
     TransportCapabilityDescriptor Capabilities => TransportCapabilityDescriptors.Unknown(GetType().Name);
 
+    Uri GetPublishAddress(string entityName) => new($"exchange:{entityName}");
+
+    Uri GetTemporaryEndpointAddress(string endpointName) =>
+        new($"exchange:{endpointName}?durable=false&autodelete=true");
+
     Task<ISendTransport> GetSendTransport(Uri address, CancellationToken cancellationToken = default);
 
     Task<IReceiveTransport> CreateReceiveTransport(